+++ data/lexicon/index.js
@@ -147,4 +147,4 @@ for (let i = 0; i < data.length; i++) {
 
 export default lex
 // console.log(Object.keys(lex).length);
-// console.log(lex[`will want`])
+// console.log(lex[`grown`])